spearmint: Minor BSPC changes
Variables holding offsetof should be size_t.
Add missing "break" so aasinfo doesn't say "don't know what to do"
Make "float.h" include be <float.h>, cppcheck complained it wasn't found.

+++ code/bspc/brushbsp.c
@@ -377,7 +377,7 @@ node_t *AllocNode (void)
 bspbrush_t *AllocBrush (int numsides)
 {
 	bspbrush_t	*bb;
-	int			c;
+	size_t		c;
 
 	c = (size_t)&(((bspbrush_t *)0)->sides[numsides]);
 	bb = GetMemory(c);
@@ -439,7 +439,7 @@ void FreeBrushList (bspbrush_t *brushes)
 bspbrush_t *CopyBrush (bspbrush_t *brush)
 {
 	bspbrush_t *newbrush;
-	int			size;
+	size_t		size;
 	int			i;
 
 	size = (size_t)&(((bspbrush_t *)0)->sides[brush->numsides]);
